rename PageLockGuard to ConflictLockGuard
